Update unit tests to test out-handler setup logic
